+++ apps/common/components/Apps.tsx
@@ -43,7 +43,7 @@ export const APPS: {[key in AppName]: TApp} = {
 		href: '/vaults-v3',
 		menu: VAULTS_V3_MENU,
 		manifest: vaultsManifest,
-		isDisabled: true,
+		isDisabled: false,
 		icon: (
 			<LogoYearn
 				className={'h-8 w-8'}